Refactor module capture / upload to Lambda.
@lambda now picks up `using` keyword inside function definition.
For each required module that is not already in the base runtime image
the .jl files are located in the module path and added to the .ZIP for
the lambda.
The lambda deployed and run in a one-time precompilation mode.
The precomile cache (*.ji) is dowloaded and added to the .ZIP file.
The lambda is the re-deployed with the updated .ZIP.